+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Change log
 
 - Deprecated `Krizalys\Onedrive\ConflictBehavior` in favor of
   `Krizalys\Onedrive\Constant\ConflictBehavior`.
+- Refactored functional tests.
 
 [2.4.0] - 2019-09-01
 --------------------