+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 Change log
 ==========
 
-[Unreleased][unreleased]
-------------------------
+[2.4.1] - 2019-09-08
+--------------------
 
 ### Added
 
@@ -256,7 +256,8 @@ as the official PHP Docker image.
 
 - Support for SSL.
 
-[unreleased]:    https://github.com/krizalys/onedrive-php-sdk/compare/2.4.0...HEAD
+[unreleased]:    https://github.com/krizalys/onedrive-php-sdk/compare/2.4.1...HEAD
+[2.4.1]:         https://github.com/krizalys/onedrive-php-sdk/compare/2.4.0...2.4.1
 [2.4.0]:         https://github.com/krizalys/onedrive-php-sdk/compare/2.3.0...2.4.0
 [2.3.0]:         https://github.com/krizalys/onedrive-php-sdk/compare/2.2.0...2.3.0
 [2.2.0]:         https://github.com/krizalys/onedrive-php-sdk/compare/2.1.2...2.2.0